# removeSong  - 3/3 tests + 2 assistance functions
1)getSongLocationByID -
The function gets a song id, if the song is in player it returns its index from the list, if not return undefind

2)removeSongFromPlaylist -
The function gets a playlist index and song id, it checks if it exists and if so removes it.

3) removeSong -
The function first checks if the song exists in the Player with the help of the function getSongLocationByID, if so it will remove it with the help of the returned index.
Secondly it goes through all the playlists and for each of them activates the function removeSongFromPlaylist, which removes the song if necessary.