# searchByQuery - 3/3 tests  + 2 assistance function
1) SearchSongByQuery-
The function gets a song and a query. The function converts all those mentioned to Lower-case and checks whether the artist title or album contains the query - then returns true.

2)SearchPlaylistByQuery - The function gets a playlist and a query. The function converts all those mentioned to Lower-case and checks if the name contains the query - then returns true.

3)searchByQuery -
The function receives a query. The function creates an object that contains the answers.
First it checks with the help of functions (1) and (2) whether there are relevant songs and playlists and inserts them into the object.
Finally it sorts the arrays created alphabetically songs by title and playlists by name.