# searchByDuration - 2/2 tests
The function converts the duration to a "seconds" format. It then checks for all the songs whether there is a (absolute) difference smaller between the times than the smallest difference. If so the song will be the closest object.
Then continues to the playlists and does the same.
Finally returns the object that is kept as close as possible.